refactor: 검증된 로또 번호(지난주 당첨 번호)를 받을 수 있도록 리팩토링
- 객체지향 생활체조 규칙3. 모든 원시값과 문자열을 포장
