+++ src/test/java/lotto/controller/LottoGameTest.java
@@ -1,6 +1,7 @@
-package lotto.domain;
+package lotto.controller;
 
 import lotto.controller.LottoGame;
+import lotto.domain.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -30,7 +31,7 @@ class LottoGameTest {
     @ParameterizedTest
     @CsvSource(value = {"1,2,3,4,5,6:6:FIRST", "1,2,3,4,5,7:5:SECOND", "1,2,3,4,7,8:4:THIRD", "1,2,3,7,8,9:3:FOURTH"}, delimiter = ':')
     public void 번호_맞춘_개수에_대한_등수_확인(String lottoNumbers, int matchingCount, Ranking expected) {
-        String lastWinningLotto = "1,2,3,4,5,6";
+        Lotto lastWinningLotto = new Lotto("1,2,3,4,5,6");
         PurchasedLotto purchasedLotto = new PurchasedLotto(Arrays.asList(new Lotto(lottoNumbers)));
         LottoResult lottoResult = game.matchLottoNumbers(purchasedLotto, lastWinningLotto);
         assertThat(lottoResult.findRankings(matchingCount).get(0)).isEqualTo(expected);
